Change Long to BigInteger of ID and ISBN

+++ src/main/java/pl/coderslab/library/Model/Book.java
@@ -3,6 +3,7 @@ package pl.coderslab.library.Model;
 import org.hibernate.validator.constraints.Length;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "book")
@@ -11,7 +12,7 @@ public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id")
-    private Long id;
+    private BigInteger id;
 
     @Column(name="title")
     private String title;
@@ -24,12 +25,12 @@ public class Book {
 
     @Column(name="isbn")
     @Length(min=13, max=13)
-    private Long isbn;
+    private BigInteger isbn;
 
     public Book() {
     }
 
-    public Book(String title, String author, String publisher, Long isbn) {
+    public Book(String title, String author, String publisher, BigInteger isbn) {
         this.title = title;
         this.author = author;
         this.publisher = publisher;
